Reworked XML response generation to better reuse allocated memory
Generating the XML response used an freshly created string for each
response. This resulted on platforms with little memory to out of memory
issues. With this commit the memory of the final storage is used to
generate this string. This reduces the number of allocations
dramatically increasing speed and reduces memory consumption.