Rework and Clean Composite FB Data Structures
As follow up to the work for the fb interface spec (see Issue #392) the
data structures for CFBs have been upgraded to spans and std::arrays.
Furthermore not needed data is now removed (e.g., internal fb and fb
params).

All main CFBs have been updated.

The arrow head module is only partly upgrade. It is in a very bad stage.
But as it is only generated code the better solution is to regenerate
the whole module after the full clean-up is done.

Releated to: https://github.com/eclipse-4diac/4diac-forte/issues/392